# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Batoilogic G01",
        "version": "1.0.0"
    },
    "paths": {
        "/api/order": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Gets all orders",
                "description": "Obtains all orders that are stored in the database",
                "operationId": "getOrders",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/order"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Add an order",
                "description": "Store a new order in the database",
                "operationId": "postOrder",
                "requestBody": {
                    "description": "Posting an order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/orderResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Show an order",
                "description": "Show an order stored in the database",
                "operationId": "showOrder",
                "requestBody": {
                    "description": "Posting an order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/orderResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "orders"
                ],
                "summary": "Edits an order",
                "description": "Overwrites an order stored in the database",
                "operationId": "updateOrder",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/orderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "orders"
                ],
                "summary": "Deletes an order",
                "description": "Deletes an order stored in the database",
                "operationId": "deletesOrder",
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/provider": {
            "get": {
                "tags": [
                    "providers"
                ],
                "summary": "Gets all providers",
                "description": "Obtains all providers that are stored in the database",
                "operationId": "getProviders",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/providerResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "providers"
                ],
                "summary": "Add a provider",
                "description": "Store a new provider in the database",
                "operationId": "postprovider",
                "requestBody": {
                    "description": "Posting a provider",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/providerResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/provider/{id}": {
            "get": {
                "tags": [
                    "providers"
                ],
                "summary": "Show a provider",
                "description": "Show a provider stored in the database",
                "operationId": "showProvider",
                "requestBody": {
                    "description": "Posting a provider",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/providerResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "providers"
                ],
                "summary": "Edits a provider",
                "description": "Overwrites a provider stored in the database",
                "operationId": "updateProvider",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/providerResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "providers"
                ],
                "summary": "Deletes an provider",
                "description": "Deletes a provider stored in the database",
                "operationId": "deletesprovider",
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The user has not been authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "orderResource": {
                "title": "orderResource",
                "description": "recurso del proyecto",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data Wraper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/order"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "orderResource"
                }
            },
            "providerResource": {
                "title": "providerResource",
                "description": "recurso del proyecto, proveedor",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data Wraper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/provider"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "providerResource"
                }
            },
            "order": {
                "required": [
                    "id,title"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "dealer_id": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "6"
                    },
                    "client_id": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "5"
                    },
                    "state_id": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Date expected for the package to arrive",
                        "type": "date",
                        "readOnly": "true",
                        "example": "2000-04-12"
                    },
                    "order": {
                        "description": "Order that determines wich order will be delivered first",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "6"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "order"
                }
            },
            "provider": {
                "required": [
                    "id,name,email,phone"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "6"
                    },
                    "logo": {
                        "description": "Photo that is assigned as the provider's logo",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "email": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "phone": {
                        "description": "Phone assigned to the provider",
                        "type": "string",
                        "readOnly": "true",
                        "example": "606055814"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "provider"
                }
            }
        },
        "responses": {
            "403": {
                "description": "Forbidden"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}